vision/requirements.txt
tensorrt
pycuda
importlib
numpy
json
robotpy-cscore
pynetworktables